Add relocated type definition files
