+++ app/features/juz/[juzId]/page.tsx
@@ -7,7 +7,8 @@ import { Verse } from '@/app/features/surah/[surahId]/components/Verse';
 import { SettingsSidebar } from '@/app/features/surah/[surahId]/components/SettingsSidebar';
 import { TranslationPanel } from '@/app/features/surah/[surahId]/components/TranslationPanel';
 import { WordLanguagePanel } from '@/app/features/surah/[surahId]/components/WordLanguagePanel';
-import { Verse as VerseType, TranslationResource, Juz } from '@/types';
+import { Verse as VerseType } from '@/types';
+import type { TranslationResource, Juz } from '@/lib/api';
 import { getTranslations, getWordTranslations, getVersesByJuz, getJuz } from '@/lib/api';
 import { LANGUAGE_CODES } from '@/lib/text/languageCodes';
 import type { LanguageCode } from '@/lib/text/languageCodes';
