--- types/chapter.ts
@@ -1,8 +1,0 @@
-export interface Chapter {
-  id: number;
-  name_simple: string;
-  name_arabic: string;
-  revelation_place: string;
-  verses_count: number;
-  pages?: [number, number];
-}
