--- types/tafsir.ts
@@ -1,5 +1,0 @@
-export interface TafsirResource {
-  id: number;
-  name: string;
-  language_name: string;
-}
