--- types/translation.ts
@@ -1,5 +1,0 @@
-export interface TranslationResource {
-  id: number;
-  name: string;
-  language_name: string;
-}
